Refactor stubs
