+++ lib/error_reporter.js
@@ -1,23 +1,6 @@
 module.exports = function(pkg, env) {
   if (!env.ERROR_REPORTER_URL) {
-    var noop = function() {};
-    var emptyMiddleware = function (a, b, next) { next(); };
-    var emptyClient = {
-      isActive: false,
-      captureException: noop,
-      patchGlobal: noop,
-      hapi: {
-        plugin: {
-          register: emptyMiddleware
-        }
-      },
-      express: {
-        requestHandler: emptyMiddleware,
-        errorHandler: emptyMiddleware
-      },
-    };
-    emptyClient.hapi.plugin.register.attributes = { pkg: require('../package.json') };
-    return emptyClient;
+    return require('./stubs').errorReporter;
   }
 
   var raven = require('raven');
