+++ lib/metrics.js
@@ -1,13 +1,6 @@
 module.exports = function(pkg, env) {
   if (!env.METRICS_API_KEY) {
-    var noop = function() {};
-    return {
-      isActive: false,
-      gauge: noop,
-      increment: noop,
-      histogram: noop,
-      flush: noop
-    };
+    return require('./stubs').metrics;
   }
 
   process.env.DATADOG_API_KEY = env.METRICS_API_KEY;
